Refaktorert struktur, lagt til oppgaver
