+++ src/TaskWrapper.js
@@ -1,0 +1,7 @@
+export const TaskWrapper = ({ children }) => {
+  return (
+    <div className="App">
+      <header className="App-header">{children}</header>
+    </div>
+  );
+};
