--- src/components/login/Greeting.jsx
@@ -1,6 +1,0 @@
-import { useCurrentUser } from "../../contexts/CurrentUser";
-
-export const Greeting = () => {
-  const { currentUser } = useCurrentUser();
-  return <p>You logged in as {currentUser.name}.</p>;
-};
