--- src/fasit/Fasit3.js
@@ -1,3 +1,0 @@
-export function Fasit3() {
-  return <p>Fasit 3</p>;
-}
