+++ src/fasit/fasit4/Fasit4.jsx
@@ -1,0 +1,15 @@
+import { DarkmodeToggle } from "./components/DarkmodeToggle";
+import { WelcomePanel } from "./components/WelcomePanel";
+import { DarkModeProvider } from "./contexts/DarkMode";
+import { CurrentUserProvider } from "./contexts/CurrentUser";
+
+export function Fasit4() {
+  return (
+    <DarkModeProvider>
+      <CurrentUserProvider>
+        <WelcomePanel />
+        <DarkmodeToggle />
+      </CurrentUserProvider>
+    </DarkModeProvider>
+  );
+}
