+++ src/fasit/fasit4/components/Greeting.jsx
@@ -1,0 +1,6 @@
+import { useCurrentUser } from "../contexts/CurrentUser";
+
+export const Greeting = () => {
+  const { currentUser } = useCurrentUser();
+  return <p>You logged in as {currentUser.name}.</p>;
+};
