+++ src/fasit/index.js
@@ -1,7 +1,7 @@
 import { Fasit1 } from "./Fasit1";
 import { Fasit2 } from "./Fasit2";
-import { Fasit3 } from "./Fasit3";
-import { Fasit4 } from "./Fasit4";
-import { Fasit5 } from "./Fasit5";
+import { Fasit3 } from "./fasit3/Fasit3";
+import { Fasit4 } from "./fasit4/Fasit4";
+import { Fasit5 } from "./fasit5/Fasit5";
 
 export { Fasit1, Fasit2, Fasit3, Fasit4, Fasit5 };
