--- src/oppgaver/Oppgave3.js
@@ -1,4 +1,0 @@
-export function Oppgave3() {
-  // Her skal du lage
-  return <p>Oppgave 3</p>;
-}
