--- src/oppgaver/Oppgave4.js
@@ -1,8 +1,0 @@
-export function Oppgave4() {
-  // Her skal du lage
-  return (
-    <div>
-      <p>Oppgave 4</p>
-    </div>
-  );
-}
