--- src/oppgaver/Oppgave5.js
@@ -1,7 +1,0 @@
-export function Oppgave5() {
-  return (
-    <div>
-      <p>Oppgave 5</p>
-    </div>
-  );
-}
