+++ src/oppgaver/index.js
@@ -1,7 +1,7 @@
 import { Oppgave1 } from "./Oppgave1";
 import { Oppgave2 } from "./Oppgave2";
-import { Oppgave3 } from "./Oppgave3";
-import { Oppgave4 } from "./Oppgave4";
-import { Oppgave5 } from "./Oppgave5";
+import { Oppgave3 } from "./oppgave3/Oppgave3";
+import { Oppgave4 } from "./oppgave4/Oppgave4";
+import { Oppgave5 } from "./oppgave5/Oppgave5";
 
 export { Oppgave1, Oppgave2, Oppgave3, Oppgave4, Oppgave5 };
