+++ src/oppgaver/oppgave4/components/Panel.jsx
@@ -1,0 +1,14 @@
+import { useDarkMode } from "../contexts/DarkMode";
+
+export const Panel = ({ title, children }) => {
+  // 9. Erstatt linjen under med darkMode fra useDarkMode hooken
+  const darkMode = null;
+  const theme = darkMode ? "dark" : "light";
+  const className = "panel-" + theme;
+  return (
+    <section className={className}>
+      <h1>{title}</h1>
+      {children}
+    </section>
+  );
+};
