+++ src/oppgaver/oppgave4/contexts/DarkMode.js
@@ -1,0 +1,14 @@
+import React, { createContext, useContext, useState } from "react";
+
+// 1. Lag DarkMode slik du gjorde i oppgave 3
+// DarkModeContext, DarkModeProvider og useDarkMode
+
+// Gå så videre til ./context/CurrentUser.js
+
+export const DarkModeProvider = ({ children }) => {
+  return children;
+};
+
+export const useDarkMode = () => {
+  return null;
+};
