+++ src/oppgaver/oppgave5/components/TodoList.jsx
@@ -1,0 +1,18 @@
+import React from "react";
+import { useTodos } from "../context/Todos";
+import { TodoItem } from "./TodoItem";
+
+export const TodoList = () => {
+  // 10. Erstatt todos med todos fra useTodos hooken
+  // Gå til ./components/TodoItem.jsx
+  const todos = [];
+  return (
+    <ul>
+      {todos.map((todo) => (
+        <li key={todo.id}>
+          <TodoItem todo={todo} />
+        </li>
+      ))}
+    </ul>
+  );
+};
